+++ packages/provider-elasticsearch/CHANGELOG.md
@@ -55,6 +55,10 @@
     * Kill `caseSensitive`, `anyOrder`, `maxWords` which weren't relevant anyway
     * Kill `fieldMode` - facet will now always use the notAnalyzed field if available
     * Kill `cardinality`, which was a proxy for precision_threshold
+  * [Number]
+    * Don't run stats/percentiles on number unless part of `findBestRange`, improving performance for the most common number filters cases
+    * Context no longer includes statistical or percentile info, only bestRange results (and no longer runs an extra extended stats call!)
+    
   * [Terms_Stats]
     * Killed useless `caseSensitive` flag
   * [Text]